6/4

+++ UI/Areas/Admin/Controllers/Checkout_CustomerController.cs
@@ -14,14 +14,33 @@ namespace UI.Areas.Admin.Controllers
         ServiceRepository service = new ServiceRepository();
 
         // GET: Admin/Checkout_Customer
-        public ActionResult Index()
+        public ActionResult Index(string seachby, int? timkiemtim)
         {
+            
+           
+           
+
+            
+            if (seachby == "id")
+            {
+                HttpResponseMessage responseMessage1 = service.GetResponse("api/Checkout_Customer/GetListcustomerById/" + timkiemtim);
+                responseMessage1.EnsureSuccessStatusCode();
+                List<DTO_Checkout_Customer> dtocustomer = responseMessage1.Content.ReadAsAsync<List<DTO_Checkout_Customer>>().Result;
+                return View(dtocustomer);
+            }
 
 
-            HttpResponseMessage responseMessage = service.GetResponse("api/Checkout_Customer/getallcustomer");
-            responseMessage.EnsureSuccessStatusCode();
-            List<DTO_Checkout_Customer> DTO_Checkout_Customers = responseMessage.Content.ReadAsAsync<List<DTO_Checkout_Customer>>().Result;
-            return View(DTO_Checkout_Customers);
+            else
+            {
+                HttpResponseMessage responseMessage = service.GetResponse("api/Checkout_Customer/getallcustomer");
+                responseMessage.EnsureSuccessStatusCode();
+                List<DTO_Checkout_Customer> DTO_Checkout_Customers = responseMessage.Content.ReadAsAsync<List<DTO_Checkout_Customer>>().Result;
+                return View(DTO_Checkout_Customers);
+                
+            }
+                   
+
+
         }
         [HttpGet]
         public ActionResult Edit(int id)
@@ -37,7 +56,7 @@ namespace UI.Areas.Admin.Controllers
         public ActionResult Details(int id)
         {
             ServiceRepository service = new ServiceRepository();
-            HttpResponseMessage responseMessage = service.GetResponse("api/Checkout_Customer/GetAccountById/" + id);
+            HttpResponseMessage responseMessage = service.GetResponse("api/Checkout_Customer/GetCustomerById/" + id);
             responseMessage.EnsureSuccessStatusCode();
             DTO_Checkout_Customer dtocustomer = responseMessage.Content.ReadAsAsync<DTO_Checkout_Customer>().Result;
 
@@ -84,29 +103,43 @@ namespace UI.Areas.Admin.Controllers
         //public ActionResult Delete(int id)
         //{
         //    ServiceRepository service = new ServiceRepository();
-        //    HttpResponseMessage responseMessage = service.GetResponse("api/Checkout_Customer/GetAccountById/" + id);
+        //    HttpResponseMessage responseMessage = service.GetResponse("api/Checkout_Customer/GetCustomerById/" + id);
         //    responseMessage.EnsureSuccessStatusCode();
         //    DTO_Checkout_Customer dtocustomer = responseMessage.Content.ReadAsAsync<DTO_Checkout_Customer>().Result;
 
         //    return View(dtocustomer);
         //}
 
-        public ActionResult Delete(int id)
+        public ActionResult Delete(int id, DTO_Checkout_Customer dTO_Checkout_Customer)
         {
             try
             {
                 // TODO: Add delete logic here
+                
+                HttpResponseMessage responseMessage = service.GetResponse("api/Checkout_Order/GetOrderByIdKH/" + id);
+                responseMessage.EnsureSuccessStatusCode();
+                DTO_Checkout_Order dtoOrder = responseMessage.Content.ReadAsAsync<DTO_Checkout_Order>().Result;
+                if (dtoOrder == null)
+                {
+                    HttpResponseMessage response = service.DeleteResponse("api/Checkout_Customer/Deletecustomer/" + id);
+                    response.EnsureSuccessStatusCode();
 
 
-                HttpResponseMessage response = service.DeleteResponse("api/Checkout_Customer/Deletecustomer/" + id);
-                response.EnsureSuccessStatusCode();
+                    return RedirectToAction("Index");
+                }
+                else
+                {
+                    ViewBag.Mess = "Không thể xóa vì khách hàng đang còn đơn đặt hàng";
+                    return View(dTO_Checkout_Customer);
+                }
+               
 
-
-                return RedirectToAction("Index");
+               
             }
             catch
             {
-                return View();
+                ViewBag.Mess = "Có lỗi ngoài ý muốn, vui lòng kiểm tra lại";
+                return View(dTO_Checkout_Customer);
             }
         }
     }
